fix: Распределение по ЕТ
