+++ zheleznyakov.mark/T2/iofmtguard.hpp
@@ -1,0 +1,16 @@
+#include <sstream>
+
+namespace zheleznyakov
+{
+  class iofmtguard
+  {
+  public:
+    iofmtguard(std::basic_ios< char > &s);
+    ~iofmtguard();
+  private:
+    std::basic_ios< char > &s_;
+    char fill_;
+    std::streamsize precision_;
+    std::basic_ios< char >::fmtflags fmt_;
+  };
+};
