+++ zheleznyakov.mark/T2/structures.cpp
@@ -1,8 +1,8 @@
+#include "iofmtguard.hpp"
 #include "structures.hpp"
 
 namespace zheleznyakov
 {
-
   using sep = DelimiterIO;
   using label = LabelIO;
   using rl = RealIO;
@@ -126,19 +126,5 @@ namespace zheleznyakov
     out << "key3 \"" << src.key3 << "\"";
     out << ":)";
     return out;
-  }
-
-  iofmtguard::iofmtguard(std::basic_ios<char> &s):
-    s_(s),
-    fill_(s.fill()),
-    precision_(s.precision()),
-    fmt_(s.flags())
-  {}
-
-  iofmtguard::~iofmtguard()
-  {
-    s_.fill(fill_);
-    s_.precision(precision_);
-    s_.flags(fmt_);
   }
 }
